Working on thread safe status messages
Virtually all the camera and scope routines use some GUI messages
that will have to be removed before they can be threaded.

This commit cleans up the status message piece.  It seems overly
complicated, but it was the simplest I could come up with and it hides
the complexity from the user.

Messages bound for the status bar are sent to the frame, and it decides
whether to display them immediately or to send them as an event to the
main thread based on whether the request comes from the main thread or not.

In the event handler, messages for all but pane 1 are sent immediately.
Pane 1 messages can be accompanined by a "time to display" because
otherwise the guider messsages were being processed with weird timing,
since they were being proceesed at some later time by the frame.