Changes to assist enabling threading
AO support is going to require the ability to move the main scope
to recenter the AO while while taking images and guiding with the
AO.  In order to have the scope moving while we are imaging, we
need do the movement on a background thread.

This commit adds a proxy class for wxMessageBox, which is one of
the main GUI functions called by the camera capture routines and
the scope guide routines. The Mount class and the Camera class
are now subclasses of the wxMessageBoxProxy class, so that any
wxMessageBox calls that are not already running on the main
thread are redirected to the main thread.

It also adds a wrapper to MyApp->Yield() that only yields
if called from the main thread.

These changes, along with the already comitted StatusText changes
will allow most of the camera and scopes to be run on the worker
thread.  Cameras that use the Windows video interface will require
more thought.