Remove some asserts that unexpectly fail
Much to my surprise, wxWidgets seems to have more than one thread,
and sometimes it runs the message loop ona thread that it does not
consider to be the main thread, so occaisionally one of these
asserts would fail. It was not on our background thread, which was
what I was trying to check.