Fix a few minor memory leaks
I got Visual Leak Detector installed, and it immediately detected
a few memory leaks.  None were in a loop, so they were minor
but memory leaks are sloppy programming, so I'm fixing them.

The first involved allocating but never freeing the memory pointed
to by pConfig, and the fix was to change pConfig into PhdConfig
since the config structure never should have been a pointer.

The second was an image allocated in the guider constructor but
not freed in the destructor.

The third was the global pCamera not get destructed anywhere