Work on fixing Transform problems
There were lots of issues the transforms based upon
using two angles and two cos() instead of a cos() and
a sin().

I think this fixes them.  I have not had an opportunity
to test them under the stars with an AO, but as far as
I can tell using the simulator it works at least as well
as it did before.