Changes to the threading model
There are now three queues per thread, so that it won't service
exposure requests while move requests are pending.

Also added a SynchronousOnly() method to mounts if they cannot be
run on the secondary thread. As far as I know, only the onboard
camera mounts fit this category, since we cannot move that
mount and image at the same time (at least not for sbig...)