Yet another round of transform/untransform changes
I added some additional tests, and these changes seem to pass
the new tests as well as all the old (something the last
changes did not do).

I don't want so say anything stupid, like "this is the end of
the transform changes", but if it isn't, I might have to quit
my volunteer phd coder job.

:-)