Rewrite the transform routine (yet another time)
We now track the xAngle, and an "error offset" for the
yAngle, and do the transforms based on
cos(xAngle)/sin(xAngle+yError) instead of
cos(xAngle)/cos(yAngle). There were sign problems using
cosine to emulate sin, and while it must be possible to
get the cosine based routines to work, but I never managed
to.

There was an error in the stepguider calibration
routine that caused the stepguider UP/DOWN calibration to
be inverted that might have been the cause of all the reversed
direction issues, so it might be possible to back out a bunch
of these chagnes, but this version seems to work, and
I've "fixed" it too many times to do so again when it
ain't broke.

I even updated the comments to reflect the changes I made,
and this checkin also contains an improved TestTransform()
that was very useful in finally figuring out what was going
on.