persist more state across sessions (Issue #60):
  exposure time
  gamma slider value
  graph x scale
  graph y scale
  target x scale
  target y scale
  camera enable subframe

Also, saves the calibration data for the mount and the ao whenever they are calibrated. This is
useful for debugging PHD2 stability issues when you need to exit phd2 and want to get going again without
waiting for calibration.  To restore the calibration, do Tools=>Enter calibration data and you will be
prompted to load the prior calibration data. If you say "no", it will go on to let you enter calibration data
normally.