Star profile layout tweaks (#1163)
* Star Profile: bugfix - make sure to initialize state variables.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

* Star profile: fix loss of precision due to integer division resulting
inaccurate position of the small red crosshairs. Also, slightly increase its
size to make it more prominent on 4K displays.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

* Star Profile: change order of rendering star image and star metrics to make the text visible.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

* Star Profile: fixed the scaling issue in star rendering within the window. Previously,
without applying the scaleFactor, the star size could become disproportionately large,
particularly when PHD2 was running in a maximized state.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

* Star Profile: completed the implementation of code (previously marked as TODO) for fine-tuning
the scaling factor. This adjustment is based on a precise calculation of the text width, assuming
linear font scaling and uniform text font width across all decimal digits.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

* Star Profile: modified to prevent displaying the star position text
when it would overlap with the star image.

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>

---------

Signed-off-by: Leo Shatz <leonid.shatz@gmail.com>