Windows build: improvement to DLL copying
Copy the Windows DLL with a single call to `cmake -E` rather than one call per
DLL. This slightly speeds up the build for a better experience in iterative
developement on Windows.

Also moves the DLL copying to POST_BUILD rather then PRE_BUILD as there's
no point in copying the DLLs if there's going to be a compile error
during iterative development.