Compatibility fixes for building a 64-bit Windows executable
On 64-bit windows sizeof(long) is 4 and sizeof(void*) is 8 so
code that casts pointers to long or vice-versa needs to be fixed.

In most cases we can replace long with intptr_t.

That fix won't work for wxWidgets GetExtraLong/SetExtraLong and
a pointer cannot be passed through those methods. Instead we can
pass the pointer in pieces in the int and long message fields.