Rename FullSize to FrameSize
FrameSize is a better name since it represents the guide frame image
size which may or not be the same size as the camera's full frame size.

There is one camera (Altair) where it is already the case that the frame size
may be less than the camera sensor size.  We plan on extending this to
other cameras so that it is possible to limit the frame size to be less than
the full camera frame size.