Dark subtraction for limited frame ROI
When a limited frame ROI is in effect, PHD2 can now properly do
dark subtraction and bad-pixel correction.

Dark frame captures now always ignore the frame limit and capture full frames.

When it is time to subtract the dark frame from the light frame,
the appropriate region of the dark frame is subtracted from the light
taking the frame limit into account.

Likewise, for a bad-pixel map, the frame limit is taken into account
when mapping the pixel position in the bad pixel map to the offset
pixel position in the light frame.

Testing:
  - manually tested a dark library and bad-pixel map
     - varied the limit frame (no limit frame, limit frame in various sizes and offsets)
     - tested with and without subframes