code cleanup: use local variables for pCamera->Binning
In preparation for adding software binning, getting the current binning
value is going to be a method call, not longer a direct member variable
access.

This PR will simplify the following PRs by storing the binning value
in a local variable rather than repeatedly accessing pCamera->Binning.

Other variables in the vicinity of the change--like focal length and
pixel scale--are also stored in local variables to make the code slightly
more concise and readable.