dark subtraction minor code refactor (#1381)
Generalize the code to allow subtracting a region of the dark frame
when there is a limiting ROI in effect for the light frame.

No functional changes in this PR (except for some improved logging
when the light and dark frames are incompatible).
Just cleanup and generalizing for use in following PR.