GuideCamera HasBayer property (#1404)
Move the "has Bayer" property from GuideCamera child classes up to the parent class.
This will allow moving the image post-processing up to the parent class as
well in a following PR.

No functional changes, just adding the property to the parent class
and replacing the child class member variables with the member from
the parent class.